Collapse multi ui fanout duplication
