+++ test/llm/ui/test_multi_ui_fanout.py
@@ -1,9 +1,23 @@
-import asyncio
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
 from zrb.llm.ui.multi_ui import MultiUI
+
+
+@pytest.fixture
+def mock_child_ui():
+    """Create a mock child UI for testing."""
+    ui = MagicMock()
+    ui.append_to_output = MagicMock()
+    ui.ask_user = AsyncMock(return_value="y")
+    ui.tool_call_handler = MagicMock()
+    ui.tool_call_handler.check_policies = AsyncMock(return_value=None)
+    ui.tool_call_handler.handle = AsyncMock(return_value=MagicMock(approved=True))
+    ui.plan_mode_active = False
+    ui.snapshot_manager = None
+    ui.history_manager = None
+    return ui
 
 
 @pytest.fixture
@@ -15,12 +29,8 @@ def child_ui_1():
     ui.run_interactive_command = AsyncMock(return_value=0)
     ui.run_async = AsyncMock(return_value="done 1")
     ui.cancel_pending_confirmations = MagicMock()
-    # Mock some expected properties/methods that MultiUI might delegate to
     ui.tool_call_handler = MagicMock()
     ui.tool_call_handler.handle = AsyncMock(return_value="Approved 1")
-    # Explicit non-mock state so _stream_ai_response's plan-mode sync and
-    # snapshot path behave as they would with a real UI (a MagicMock would be
-    # truthy and flip the global agent-mode ContextVar, polluting other tests).
     ui.plan_mode_active = False
     ui.snapshot_manager = None
     ui.history_manager = None
@@ -44,11 +54,25 @@ def multi_ui(child_ui_1, child_ui_2):
     return MultiUI([child_ui_1, child_ui_2])
 
 
-def test_multi_ui_init(multi_ui, child_ui_1, child_ui_2):
-    assert child_ui_1.multi_ui_parent is multi_ui
-    assert child_ui_2.multi_ui_parent is multi_ui
-    # multi_ui.main_ui is a property
-    assert multi_ui.main_ui is child_ui_1
+def test_append_to_output_broadcasts(mock_child_ui):
+    """Test that append_to_output broadcasts to all child UIs."""
+    other_ui = MagicMock()
+    other_ui.append_to_output = MagicMock()
+    multi_ui = MultiUI([mock_child_ui, other_ui])
+
+    multi_ui.append_to_output("Test message")
+
+    mock_child_ui.append_to_output.assert_called_once()
+    other_ui.append_to_output.assert_called_once()
+
+
+def test_append_to_output_handles_exception(mock_child_ui):
+    """Test append_to_output handles exceptions from child UIs."""
+    mock_child_ui.append_to_output = MagicMock(side_effect=Exception("Test"))
+    multi_ui = MultiUI([mock_child_ui])
+
+    # Should not raise
+    multi_ui.append_to_output("Test message")
 
 
 def test_multi_ui_append_to_output(multi_ui, child_ui_1, child_ui_2):
@@ -291,199 +315,40 @@ def test_multi_ui_set_thinking_mirrors_to_children(multi_ui, child_ui_1, child_u
     assert child_ui_2.is_thinking is False
 
 
-@pytest.mark.asyncio
-async def test_multi_ui_stream_sets_thinking_on_children(multi_ui, child_ui_1):
-    multi_ui.append_to_output = MagicMock()
-    llm_task = MagicMock()
-    llm_task.async_run = AsyncMock(return_value="# Response")
-    llm_task.set_ui = MagicMock()
-    llm_task.tool_confirmation = MagicMock()
+class TestMultiUIReplayHistory:
+    """Tests for MultiUI broadcasting replay to child UIs."""
 
-    await multi_ui.stream_ai_response(llm_task, "Hello", [])
+    def test_replay_history_broadcasts_to_children(self):
+        """MultiUI.replay_history must call replay_history on every child."""
+        child_a = MagicMock()
+        child_b = MagicMock()
+        multi_ui = MultiUI([child_a, child_b])
 
-    # Thinking flag must be False after the run, not just during it.
-    assert multi_ui.is_thinking is False
-    assert child_ui_1.is_thinking is False
+        messages = ["m1", "m2"]
+        multi_ui.replay_history(messages)
 
+        child_a.replay_history.assert_called_once_with(messages)
+        child_b.replay_history.assert_called_once_with(messages)
 
-@pytest.mark.asyncio
-async def test_multi_ui_stream_uses_append_markdown_on_main_ui(multi_ui, child_ui_1):
-    multi_ui.append_to_output = MagicMock()
-    child_ui_1.append_markdown = MagicMock()
-    llm_task = MagicMock()
-    llm_task.async_run = AsyncMock(return_value="# Response")
-    llm_task.set_ui = MagicMock()
-    llm_task.tool_confirmation = MagicMock()
+    def test_replay_history_skips_children_without_method(self):
+        """Children missing replay_history are silently skipped."""
 
-    await multi_ui.stream_ai_response(llm_task, "Hello", [])
+        class NoReplayChild:
+            def __init__(self):
+                self._multi_ui_parent = None
 
-    # The main UI gets themed, re-wrappable markdown; other children (e.g.
-    # Telegram) get the pre-rendered text.
-    child_ui_1.append_markdown.assert_called_once_with("# Response")
+        child = NoReplayChild()
+        # Should not raise
+        MultiUI([child]).replay_history(["m1"])
 
+    def test_replay_history_swallows_child_errors(self):
+        """A child raising must not break the broadcast to other children."""
+        bad_child = MagicMock()
+        bad_child.replay_history.side_effect = RuntimeError("bad")
+        good_child = MagicMock()
 
-@pytest.mark.asyncio
-async def test_multi_ui_stream_uses_rendered_text_on_other_children(
-    multi_ui, child_ui_1, child_ui_2
-):
-    multi_ui.append_to_output = MagicMock()
-    child_ui_1.append_markdown = MagicMock()
-    # MagicMock auto-creates any attribute; remove it so hasattr() is False,
-    # matching a real chat backend (e.g. TelegramUI) that lacks append_markdown.
-    child_ui_2.append_markdown = None
-    llm_task = MagicMock()
-    llm_task.async_run = AsyncMock(return_value="# Response")
-    llm_task.set_ui = MagicMock()
-    llm_task.tool_confirmation = MagicMock()
+        multi_ui = MultiUI([bad_child, good_child])
+        # Should not raise even though bad_child throws
+        multi_ui.replay_history(["m1"])
 
-    await multi_ui.stream_ai_response(llm_task, "Hello", [])
-
-    # child_ui_2 has no append_markdown → gets rendered text with end="".
-    child_ui_2.append_to_output.assert_called()
-    args = child_ui_2.append_to_output.call_args
-    assert args.kwargs.get("end") == ""
-
-
-@pytest.mark.asyncio
-async def test_multi_ui_stream_takes_snapshot_before_run(multi_ui, child_ui_1):
-    multi_ui.append_to_output = MagicMock()
-    snapshot_manager = MagicMock()
-    snapshot_manager.take_snapshot = AsyncMock()
-    child_ui_1.snapshot_manager = snapshot_manager
-    child_ui_1.history_manager = MagicMock()
-    child_ui_1.history_manager.load = MagicMock(return_value=["msg1"])
-    child_ui_1.conversation_session_name = "my-session"
-
-    llm_task = MagicMock()
-    llm_task.async_run = AsyncMock(return_value="# Response")
-    llm_task.set_ui = MagicMock()
-    llm_task.tool_confirmation = MagicMock()
-
-    await multi_ui.stream_ai_response(llm_task, "Hello", [])
-
-    snapshot_manager.take_snapshot.assert_called_once()
-    kwargs = snapshot_manager.take_snapshot.call_args.kwargs
-    assert kwargs.get("message_count") == 1
-
-
-@pytest.mark.asyncio
-async def test_multi_ui_stream_syncs_plan_mode(multi_ui, child_ui_1):
-    from zrb.llm.permission.state import (
-        AgentMode,
-        get_current_agent_mode,
-        set_current_agent_mode,
-    )
-
-    multi_ui.append_to_output = MagicMock()
-    child_ui_1.plan_mode_active = True
-    llm_task = MagicMock()
-    llm_task.async_run = AsyncMock(return_value="# Response")
-    llm_task.set_ui = MagicMock()
-    llm_task.tool_confirmation = MagicMock()
-
-    try:
-        await multi_ui.stream_ai_response(llm_task, "Hello", [])
-
-        # Plan mode set on the main UI must reach the run and be read back.
-        assert get_current_agent_mode() == AgentMode.PLAN
-        assert child_ui_1.plan_mode_active is True
-    finally:
-        # Reset the module-level ContextVar so other tests don't inherit PLAN.
-        set_current_agent_mode(AgentMode.BUILD)
-
-
-@pytest.mark.asyncio
-async def test_multi_ui_stream_updates_system_info_on_children(
-    multi_ui, child_ui_1, child_ui_2
-):
-    multi_ui.append_to_output = MagicMock()
-    child_ui_1.update_system_info = AsyncMock()
-    child_ui_2.update_system_info = AsyncMock()
-    llm_task = MagicMock()
-    llm_task.async_run = AsyncMock(return_value="# Response")
-    llm_task.set_ui = MagicMock()
-    llm_task.tool_confirmation = MagicMock()
-
-    await multi_ui.stream_ai_response(llm_task, "Hello", [])
-
-    child_ui_1.update_system_info.assert_awaited_once()
-    child_ui_2.update_system_info.assert_awaited_once()
-
-
-@pytest.mark.asyncio
-async def test_multi_ui_stream_repaints_after_system_info_update(multi_ui, child_ui_1):
-    # The status bar must be repainted with fresh system info, not before it.
-    # Sequence: thinking-on repaint → system info update → final repaint.
-    multi_ui.append_to_output = MagicMock()
-    order = []
-    child_ui_1.invalidate_ui = MagicMock(side_effect=lambda: order.append("paint"))
-
-    async def _update():
-        order.append("update")
-
-    child_ui_1.update_system_info = _update
-    llm_task = MagicMock()
-    llm_task.async_run = AsyncMock(return_value="# Response")
-    llm_task.set_ui = MagicMock()
-    llm_task.tool_confirmation = MagicMock()
-
-    await multi_ui.stream_ai_response(llm_task, "Hello", [])
-
-    assert order == ["paint", "update", "paint"]
-
-
-@pytest.mark.asyncio
-async def test_multi_ui_stream_non_string_result_clears_last_output(
-    multi_ui, child_ui_1
-):
-    # A turn whose result is not a string must not leave last_output carrying
-    # the previous turn's answer.
-    multi_ui.append_to_output = MagicMock()
-    multi_ui.last_result_data = "stale"
-    llm_task = MagicMock()
-    llm_task.async_run = AsyncMock(return_value={"structured": "result"})
-    llm_task.set_ui = MagicMock()
-    llm_task.tool_confirmation = MagicMock()
-
-    await multi_ui.stream_ai_response(llm_task, "Hello", [])
-
-    assert multi_ui.last_result_data is None
-
-
-@pytest.mark.asyncio
-async def test_multi_ui_ask_user_race(multi_ui, child_ui_1, child_ui_2):
-    # Make child_ui_1 slower
-    async def slow_ask(*args, **kwargs):
-        await asyncio.sleep(0.1)
-        return "input 1"
-
-    child_ui_1.ask_user = slow_ask
-
-    # Make child_ui_2 faster
-    async def fast_ask(*args, **kwargs):
-        await asyncio.sleep(0.01)
-        return "input 2"
-
-    child_ui_2.ask_user = fast_ask
-
-    res = await multi_ui.ask_user("prompt")
-    assert res == "input 2"
-
-
-@pytest.mark.asyncio
-async def test_multi_ui_run_async(multi_ui, child_ui_1, child_ui_2):
-    multi_ui.set_llm_task(MagicMock())
-    child_ui_1.last_output = "Final Output"
-
-    res = await multi_ui.run_async()
-
-    assert res == "Final Output"
-    child_ui_1.run_async.assert_called_once()
-    child_ui_2.start_event_loop.assert_called_once()
-
-
-@pytest.mark.asyncio
-async def test_multi_ui_run_interactive_command(multi_ui, child_ui_1):
-    res = await multi_ui.run_interactive_command("ls")
-    assert res == 0
-    child_ui_1.run_interactive_command.assert_called_with("ls", shell=False)
+        good_child.replay_history.assert_called_once_with(["m1"])
