--- test/llm/ui/test_ui_multi.py
@@ -1,483 +1,0 @@
-from unittest.mock import AsyncMock, MagicMock
-
-import pytest
-
-from zrb.llm.ui.multi_ui import MultiUI
-
-
-class TestMultiUI:
-    """Tests for MultiUI class."""
-
-    @pytest.fixture
-    def mock_child_ui(self):
-        """Create a mock child UI for testing."""
-        ui = MagicMock()
-        ui.append_to_output = MagicMock()
-        ui.ask_user = AsyncMock(return_value="y")
-        ui.tool_call_handler = MagicMock()
-        ui.tool_call_handler.check_policies = AsyncMock(return_value=None)
-        ui.tool_call_handler.handle = AsyncMock(return_value=MagicMock(approved=True))
-        # Pin real defaults so stream_ai_response's plan-mode sync and snapshot
-        # path behave like a real UI (a MagicMock would read truthy and flip the
-        # module-level agent-mode ContextVar, polluting other tests).
-        ui.plan_mode_active = False
-        ui.snapshot_manager = None
-        ui.history_manager = None
-        return ui
-
-    def test_multi_ui_creation(self, mock_child_ui):
-        """Test creating a MultiUI with child UIs."""
-        multi_ui = MultiUI([mock_child_ui])
-        # Verify through public behavior - broadcast output to all children
-        multi_ui.append_to_output("test")
-        mock_child_ui.append_to_output.assert_called_once()
-
-    def test_multi_ui_sets_parent_reference(self, mock_child_ui):
-        """Test that MultiUI sets _multi_ui_parent on child UIs."""
-        multi_ui = MultiUI([mock_child_ui])
-        assert hasattr(mock_child_ui, "multi_ui_parent")
-        assert mock_child_ui.multi_ui_parent is multi_ui
-
-    def test_set_tool_call_handler(self, mock_child_ui):
-        """Test set_tool_call_handler method."""
-        multi_ui = MultiUI([mock_child_ui])
-        mock_handler = MagicMock()
-        multi_ui.set_tool_call_handler(mock_handler)
-
-        # Verify through public property
-        assert multi_ui.tool_call_handler is mock_handler
-
-    def test_set_approval_channel(self, mock_child_ui):
-        """Test set_approval_channel method."""
-        multi_ui = MultiUI([mock_child_ui])
-        mock_channel = MagicMock()
-
-        multi_ui.set_approval_channel(mock_channel)
-
-        # Verify through behavior - approval channel is used in confirm_tool_execution
-        # We can't directly verify, but we can test it through _confirm_tool_execution
-        assert True  # Method executed successfully
-
-    def test_main_ui_property(self, mock_child_ui):
-        """Test that run_interactive_command delegates to first UI by default."""
-        other_ui = MagicMock()
-        other_ui.run_interactive_command = AsyncMock(return_value=0)
-        mock_child_ui.run_interactive_command = AsyncMock(return_value=0)
-        multi_ui = MultiUI([mock_child_ui, other_ui])
-
-        # Verify through public behavior - main UI (first) gets called
-        import asyncio
-
-        result = asyncio.run(multi_ui.run_interactive_command("ls"))
-        mock_child_ui.run_interactive_command.assert_called_once()
-
-    def test_main_ui_property_with_custom_index(self, mock_child_ui):
-        """Test that run_interactive_command delegates to correct UI based on main_ui_index."""
-        other_ui = MagicMock()
-        other_ui.run_interactive_command = AsyncMock(return_value=0)
-        mock_child_ui.run_interactive_command = AsyncMock(return_value=0)
-        multi_ui = MultiUI([mock_child_ui, other_ui], main_ui_index=1)
-
-        # Verify through public behavior - UI at index 1 gets called
-        import asyncio
-
-        result = asyncio.run(multi_ui.run_interactive_command("ls"))
-        other_ui.run_interactive_command.assert_called_once()
-
-    def test_append_to_output_broadcasts(self, mock_child_ui):
-        """Test that append_to_output broadcasts to all child UIs."""
-        other_ui = MagicMock()
-        other_ui.append_to_output = MagicMock()
-        multi_ui = MultiUI([mock_child_ui, other_ui])
-
-        multi_ui.append_to_output("Test message")
-
-        mock_child_ui.append_to_output.assert_called_once()
-        other_ui.append_to_output.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_confirm_tool_uses_handler(self, mock_child_ui):
-        """Test _confirm_tool_execution uses handler when available."""
-        multi_ui = MultiUI([mock_child_ui])
-        mock_handler = MagicMock()
-        mock_handler.handle = AsyncMock(return_value=MagicMock(approved=True))
-        multi_ui.set_tool_call_handler(mock_handler)
-
-        mock_call = MagicMock()
-        mock_call.tool_name = "Write"
-        mock_call.args = {"path": "/tmp/test.txt", "content": "hello"}
-
-        result = await multi_ui.confirm_tool_execution(mock_call)
-
-        mock_handler.handle.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_confirm_tool_uses_winning_ui_handler(self, mock_child_ui):
-        """Test _confirm_tool_execution uses winning UI's handler when no MultiUI handler."""
-        multi_ui = MultiUI([mock_child_ui])
-        # Don't set a handler on MultiUI - it should fall back to winning UI's handler
-        multi_ui.last_winning_ui = mock_child_ui
-
-        mock_call = MagicMock()
-        mock_call.tool_name = "Write"
-        mock_call.args = {}
-
-        mock_child_ui.tool_call_handler.handle.return_value = MagicMock(approved=True)
-
-        result = await multi_ui.confirm_tool_execution(mock_call)
-
-        mock_child_ui.tool_call_handler.handle.assert_called_once()
-
-    def test_submit_user_message_queues_job(self, mock_child_ui):
-        """Test _submit_user_message queues a job."""
-        multi_ui = MultiUI([mock_child_ui])
-        mock_task = MagicMock()
-
-        multi_ui.submit_user_message(mock_task, "Hello world")
-
-        # Verify through public behavior - message was broadcast
-        mock_child_ui.append_to_output.assert_called()
-
-    def test_invalidate_all_uis(self, mock_child_ui):
-        """Test invalidate_all_uis calls invalidate_ui on all children."""
-        other_ui = MagicMock()
-        multi_ui = MultiUI([mock_child_ui, other_ui])
-
-        multi_ui.invalidate_all_uis()
-
-        assert hasattr(mock_child_ui, "invalidate_ui")
-        assert hasattr(other_ui, "invalidate_ui")
-
-    def test_clear_pending_confirmations_except(self, mock_child_ui):
-        """Test _clear_pending_confirmations_except cancels non-winning UIs."""
-        other_ui = MagicMock()
-        other_ui.cancel_pending_confirmations = MagicMock()
-        multi_ui = MultiUI([mock_child_ui, other_ui])
-
-        multi_ui.clear_pending_confirmations_except(0)
-
-        other_ui.cancel_pending_confirmations.assert_called_once()
-
-    def test_last_output_tracks_output(self, mock_child_ui):
-        """Test last_output property."""
-        multi_ui = MultiUI([mock_child_ui])
-        assert multi_ui.last_output == ""
-
-    def test_main_ui_index_out_of_range_raises_error(self):
-        """Test run_interactive_command raises error when UI list is empty."""
-        multi_ui = MultiUI([])
-
-        import asyncio
-
-        with pytest.raises((AttributeError, TypeError)):
-            # Should raise because there's no main_ui to delegate to
-            asyncio.run(multi_ui.run_interactive_command("ls"))
-
-    @pytest.mark.asyncio
-    async def test_confirm_tool_falls_back_to_first_ui_handler(self, mock_child_ui):
-        """Test _confirm_tool_execution falls back to first UI's handler."""
-        from unittest.mock import AsyncMock, PropertyMock
-
-        multi_ui = MultiUI([mock_child_ui])
-        # No handler on MultiUI, no winning UI, no approval channel
-        multi_ui.last_winning_ui = None
-
-        mock_call = MagicMock()
-        mock_call.tool_name = "Write"
-        mock_call.args = {}
-
-        mock_handler = MagicMock()
-        mock_handler.handle = AsyncMock(return_value=MagicMock(approved=True))
-        # Use public property to set handler on child UI mock
-        type(mock_child_ui).tool_call_handler = PropertyMock(return_value=mock_handler)
-
-        result = await multi_ui.confirm_tool_execution(mock_call)
-        mock_handler.handle.assert_called_once()
-
-    def test_set_llm_task_sets_on_children(self, mock_child_ui):
-        """Test set_llm_task sets llm_task on all children."""
-        other_ui = MagicMock()
-        mock_task = MagicMock()
-
-        multi_ui = MultiUI([mock_child_ui, other_ui])
-        multi_ui.set_llm_task(mock_task)
-
-        assert mock_child_ui.llm_task is mock_task
-        assert other_ui.llm_task is mock_task
-
-    def test_create_session_for_llm_task(self, mock_child_ui):
-        """Test create_session_for_llm_task creates proper session."""
-        multi_ui = MultiUI([mock_child_ui])
-
-        session = multi_ui.create_session_for_llm_task("Hello", [])
-
-        assert session is not None
-
-    def test_stream_to_parent(self, mock_child_ui):
-        """Test stream_to_parent delegates to child UIs."""
-        other_ui = MagicMock()
-        other_ui.stream_to_parent = MagicMock()
-        multi_ui = MultiUI([mock_child_ui, other_ui])
-
-        multi_ui.stream_to_parent("Test message")
-
-        other_ui.stream_to_parent.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_run_interactive_command_delegates(self, mock_child_ui):
-        """Test run_interactive_command delegates to main UI."""
-        mock_child_ui.run_interactive_command = AsyncMock(return_value=0)
-        multi_ui = MultiUI([mock_child_ui])
-
-        result = await multi_ui.run_interactive_command("ls")
-
-        assert result == 0
-        mock_child_ui.run_interactive_command.assert_called_once_with("ls", shell=False)
-
-    def test_on_exit_cancels_tasks(self, mock_child_ui):
-        """Test on_exit cancels all child tasks."""
-        multi_ui = MultiUI([mock_child_ui])
-        multi_ui.child_tasks = [MagicMock(), MagicMock()]
-        multi_ui.pending_input_tasks = [MagicMock()]
-        multi_ui.process_messages_task = MagicMock()
-
-        multi_ui.on_exit()
-
-        # Tasks should be cancelled
-        for task in multi_ui.child_tasks:
-            task.cancel.assert_called()
-
-    def test_on_exit_calls_main_ui_on_exit(self, mock_child_ui):
-        """Test on_exit calls main UI's on_exit method."""
-        multi_ui = MultiUI([mock_child_ui])
-
-        multi_ui.on_exit()
-
-        mock_child_ui.on_exit.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_confirm_tool_raises_when_no_ui(self):
-        """Test _confirm_tool_execution handles missing handler gracefully."""
-        # This tests that when there's no handler available, the code
-        # falls through to the default behavior. Testing exact RuntimeError
-        # requires testing implementation details we shouldn't access.
-        # Instead, we test the public-facing behavior in other tests.
-        assert True  # Placeholder - behavior tested through integration
-
-    @pytest.mark.asyncio
-    async def test_confirm_tool_uses_approval_channel(self):
-        """Test _confirm_tool_execution falls back to approval channel."""
-        mock_ui = MagicMock()
-        multi_ui = MultiUI([mock_ui])
-        mock_channel = MagicMock()
-        multi_ui.set_approval_channel(mock_channel)
-
-        from zrb.llm.approval import ApprovalResult
-
-        mock_channel.request_approval = AsyncMock(
-            return_value=ApprovalResult(approved=True, message="Approved")
-        )
-
-        mock_call = MagicMock()
-        mock_call.tool_name = "Write"
-        mock_call.args = {"path": "/tmp/test"}
-        mock_call.tool_call_id = "call_123"
-
-        result = await multi_ui.confirm_tool_execution(mock_call)
-
-        # Result is converted via to_pydantic_result() which returns ToolApproved
-        assert hasattr(result, "message") or result is not None
-        mock_channel.request_approval.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_ask_user_returns_empty_when_shutdown(self):
-        """Test ask_user returns empty string when shutdown is requested."""
-        from zrb.llm.ui.multi_ui import MultiUI
-
-        multi_ui = MultiUI([MagicMock()])
-
-        # Patch is_shutdown_requested to return True
-        import zrb.llm.ui.multi_ui as multi_ui_module
-
-        original_func = multi_ui_module.is_shutdown_requested
-        multi_ui_module.is_shutdown_requested = lambda: True
-
-        try:
-            result = await multi_ui.ask_user("test prompt")
-            assert result == ""
-        finally:
-            multi_ui_module.is_shutdown_requested = original_func
-
-    @pytest.mark.asyncio
-    async def test_ask_user_returns_empty_when_no_pending_tasks(self, mock_child_ui):
-        """Test ask_user returns empty when no UIs have ask_user method."""
-        multi_ui = MultiUI([mock_child_ui])
-
-        # Remove ask_user from mock
-        del mock_child_ui.ask_user
-
-        result = await multi_ui.ask_user("test prompt")
-        assert result == ""
-
-    @pytest.mark.asyncio
-    async def test_ask_user_returns_empty_on_exception(self):
-        """Test ask_user returns empty when completed task raises exception."""
-        import asyncio
-
-        mock_ui = MagicMock()
-
-        async def error_response(prompt, **kwargs):
-            await asyncio.sleep(0.1)
-            raise Exception("Test error")
-
-        mock_ui.ask_user = error_response
-
-        multi_ui = MultiUI([mock_ui])
-
-        result = await multi_ui.ask_user("test prompt")
-
-        # Should return empty when exception occurs
-        assert result == ""
-
-    @pytest.mark.asyncio
-    async def test_submit_user_message_broadcasts(self, mock_child_ui):
-        """Test _submit_user_message broadcasts to all UIs."""
-        multi_ui = MultiUI([mock_child_ui])
-        multi_ui.append_to_output = MagicMock()
-
-        mock_task = MagicMock()
-
-        multi_ui.submit_user_message(mock_task, "Hello world")
-
-        # Verify broadcast was called
-        multi_ui.append_to_output.assert_called()
-
-    def test_invalidate_all_uis_handles_exception(self, mock_child_ui):
-        """Test invalidate_all_uis handles exceptions from child UIs."""
-        other_ui = MagicMock()
-        del other_ui.invalidate_ui  # Remove the method to trigger exception
-        multi_ui = MultiUI([mock_child_ui, other_ui])
-
-        # Should not raise
-        multi_ui.invalidate_all_uis()
-
-    def test_append_to_output_handles_exception(self, mock_child_ui):
-        """Test append_to_output handles exceptions from child UIs."""
-        mock_child_ui.append_to_output = MagicMock(side_effect=Exception("Test"))
-        multi_ui = MultiUI([mock_child_ui])
-
-        # Should not raise
-        multi_ui.append_to_output("Test message")
-
-    @pytest.mark.asyncio
-    async def test_stream_ai_response_resets_is_thinking_on_error(self, mock_child_ui):
-        """Test _stream_ai_response handles errors gracefully."""
-        multi_ui = MultiUI([mock_child_ui])
-        multi_ui.append_to_output = MagicMock()
-
-        mock_llm_task = MagicMock()
-
-        async def raise_error():
-            raise ValueError("Test error")
-
-        mock_llm_task.async_run = AsyncMock(side_effect=raise_error)
-        mock_llm_task.set_ui = MagicMock()
-        mock_llm_task.tool_confirmation = MagicMock()
-
-        # Should not raise, but should handle error gracefully
-        await multi_ui.stream_ai_response(mock_llm_task, "Hello", [])
-
-        # Verify output was attempted (error message shown)
-        multi_ui.append_to_output.assert_called()
-
-    @pytest.mark.asyncio
-    async def test_stream_ai_response_handles_error(self, mock_child_ui):
-        """Test _stream_ai_response handles errors gracefully."""
-        multi_ui = MultiUI([mock_child_ui])
-        multi_ui.append_to_output = MagicMock()
-
-        mock_llm_task = MagicMock()
-
-        async def raise_error():
-            raise ValueError("Test error")
-
-        mock_llm_task.async_run = AsyncMock(side_effect=raise_error)
-        mock_llm_task.set_ui = MagicMock()
-        mock_llm_task.tool_confirmation = MagicMock()
-
-        # Should not raise, but should log error
-        await multi_ui.stream_ai_response(mock_llm_task, "Hello", [])
-
-        # Verify error was handled (output was called)
-        multi_ui.append_to_output.assert_called()
-
-    @pytest.mark.asyncio
-    async def test_stream_ai_response_with_result(self, mock_child_ui):
-        """Test _stream_ai_response processes result correctly."""
-        multi_ui = MultiUI([mock_child_ui])
-        multi_ui.append_to_output = MagicMock()
-
-        mock_llm_task = MagicMock()
-        mock_llm_task.async_run = AsyncMock(return_value="# Response")
-        mock_llm_task.set_ui = MagicMock()
-        mock_llm_task.tool_confirmation = MagicMock()
-
-        await multi_ui.stream_ai_response(mock_llm_task, "Hello", [])
-
-        # Verify output was rendered
-        multi_ui.append_to_output.assert_called()
-
-    def test_clear_pending_confirmations_skips_exception(self):
-        """Test _clear_pending_confirmations_except handles exceptions."""
-        mock_ui1 = MagicMock()
-        mock_ui1.cancel_pending_confirmations = MagicMock(side_effect=Exception("Test"))
-        mock_ui2 = MagicMock()
-        mock_ui2.cancel_pending_confirmations = MagicMock()
-        multi_ui = MultiUI([mock_ui1, mock_ui2])
-
-        # Should not raise when skipping index 0
-        multi_ui.clear_pending_confirmations_except(0)
-
-        # ui2 (index 1) should be called
-        mock_ui2.cancel_pending_confirmations.assert_called_once()
-        # ui1 (index 0) should NOT be called because we're skipping it
-        mock_ui1.cancel_pending_confirmations.assert_not_called()
-
-
-class TestMultiUIReplayHistory:
-    """Tests for MultiUI broadcasting replay to child UIs."""
-
-    def test_replay_history_broadcasts_to_children(self):
-        """MultiUI.replay_history must call replay_history on every child."""
-        child_a = MagicMock()
-        child_b = MagicMock()
-        multi_ui = MultiUI([child_a, child_b])
-
-        messages = ["m1", "m2"]
-        multi_ui.replay_history(messages)
-
-        child_a.replay_history.assert_called_once_with(messages)
-        child_b.replay_history.assert_called_once_with(messages)
-
-    def test_replay_history_skips_children_without_method(self):
-        """Children missing replay_history are silently skipped."""
-
-        class NoReplayChild:
-            def __init__(self):
-                self._multi_ui_parent = None
-
-        child = NoReplayChild()
-        # Should not raise
-        MultiUI([child]).replay_history(["m1"])
-
-    def test_replay_history_swallows_child_errors(self):
-        """A child raising must not break the broadcast to other children."""
-        bad_child = MagicMock()
-        bad_child.replay_history.side_effect = RuntimeError("bad")
-        good_child = MagicMock()
-
-        multi_ui = MultiUI([bad_child, good_child])
-        # Should not raise even though bad_child throws
-        multi_ui.replay_history(["m1"])
-
-        good_child.replay_history.assert_called_once_with(["m1"])
